chord/local_test: wait until no nodes have nil predecessor before RingCheck
The macOS runner on GitHub Actions consistently fails with "node X has
nil predecessor." The log it outputs suggests that it is a race
condition, however it is impossible to be a race condition becasue of
predecessorMu in both .Notify() and .getPredecessor(). To be sure, let's
wait until all the nodes have no nil predecessor until we RingCheck().