cmd/client: support for connecting TCP target
This commit introduces support for one-off or listener mode to connect
to tunneled target from the specter client. Some changes to the gateway
had to be made because of quic's EarlyConnection where the stream is not
established despite using .OpenStreamSync(). Therefore, the client needs
to poke the gateway.

Other changes in this commit:
util: move GetOutboundIP to util package
tun/client: call .Sync() before closing the config file